requirements.txt
Django==1.7.1
MySQL-python==1.2.5
Pygments==2.0.1
argparse==1.2.1
coverage==3.7.1
djangorestframework==2.4.4
djoser==0.1.0
wsgiref==0.1.2
